+++ server/lib/localObjects.py
@@ -64,5 +64,5 @@ class TimeoutSensor(_InternalSensor):
 	def __init__(self):
 		_InternalSensor.__init__(self)
 
-		# TODO description
-		self.timeoutSensors = list()
+		# a set of ids of the sensors that are timed out
+		self.timeoutSensorIds = set()